% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_210503.R
\name{compmse}
\alias{compmse}
\title{Comapre bias, var and MSE}
\usage{
compmse(ibias, ivals)
}
\arguments{
\item{ibias}{Matrix of bias's}

\item{ivals}{Matrix of var's}
}
\value{
A matrix
}
\description{
Comapre bias, var and MSE
}
