% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_210424.R
\name{coverage}
\alias{coverage}
\title{Calculate coverage probabilties}
\usage{
coverage(estimates, vars, beta, round = 3)
}
\arguments{
\item{estimates}{A matrix of estimates}

\item{vars}{A matrix of varainces}

\item{beta}{The beta under H0}

\item{round}{The decimal places for rounding}
}
\value{
95% CI coverage probabilities
}
\description{
Calculate coverage probabilties
}
