% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_211026.R
\name{coverage2}
\alias{coverage2}
\title{Compare coverage probabilities between two estimators on matched simulations}
\usage{
coverage2(estimates1, vars1, estimates2, vars2, beta, round = 3)
}
\arguments{
\item{estimates1}{Matrix of estimates for 1}

\item{vars1}{Matrix of vars for 1}

\item{estimates2}{Matrix of estimates for 2}

\item{vars2}{Matrix of vars for 2}

\item{beta}{The beta under H0}

\item{round}{The decimal places for rounding}
}
\value{
Comparison of coverage probabilities between two estimators.
}
\description{
Compare coverage probabilities between two estimators on matched simulations
}
