% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210503.R
\name{dfbetac}
\alias{dfbetac}
\title{Sum dfbeta s According to id_vector Clusters}
\usage{
dfbetac(id_vector, dfbeta)
}
\arguments{
\item{id_vector}{Input id vector}

\item{dfbeta}{dfbeta s for sandwich}
}
\value{
dfbeta by id_vector clusters
}
\description{
Sum dfbeta s According to id_vector Clusters
}
