% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210424.R
\name{meerva}
\alias{meerva}
\title{Analysis of Data with Measurement Error Using a Validation Subsample}
\usage{
meerva()
}
\value{
Describes future development of the meerva package.
}
\description{
The meerva package performs regression analyses on data
with measurement error when there is a
validation subsample.  The functional .fit program is meerva.fit. The meerva function
is intended for future development and use as a wrapper for meerva.fit.
Please help(meerva.fit).
}
\seealso{
\code{\link{meerva.fit}} , \code{\link{meerva.sim.block}} , \code{\link{meerva.sim.brn}} , \code{\link{meerva.sim.cox}} , \code{\link{meerva.sim.nrm}}
}
\author{
Walter Kremers (kremers.walter@mayo.edu)
}
