% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_211026.R
\name{myttest}
\alias{myttest}
\title{A simple summary description}
\usage{
myttest(x, beta0 = NULL)
}
\arguments{
\item{x}{A data matrix}

\item{beta0}{A vector for H0}
}
\value{
A matrix
}
\description{
A simple summary description
}
