% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_210414.R
\name{plot.meerva.sim}
\alias{plot.meerva.sim}
\title{Plot results for meerva.sim output object generated by meerva.sim.block function}
\usage{
\method{plot}{meerva.sim}(x, violin = 0, ...)
}
\arguments{
\item{x}{A meerva.sim class object}

\item{violin}{1 to produce a violin plot instead of a boxplot}

\item{...}{further arguments}
}
\value{
This displays a plot
}
\description{
Plot results for meerva.sim output object generated by meerva.sim.block function
}
