% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210503.R
\name{print.meerva}
\alias{print.meerva}
\title{Print Minimal Summary Information for a meerva Output Object}
\usage{
\method{print}{meerva}(x, alpha = 0.05, round = NA, ...)
}
\arguments{
\item{x}{A meerva class object for printing}

\item{alpha}{level for (1-alpha) confidence intervals}

\item{round}{number of decimal places to print for some outputs}

\item{...}{further arguments}
}
\value{
Print output
}
\description{
Print Minimal Summary Information for a meerva Output Object
}
