% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_211026.R
\name{print.meerva.sim}
\alias{print.meerva.sim}
\title{Print Information for a meerva.sim Simulation Study Output Object (alias for summary.meerva.fit())}
\usage{
\method{print}{meerva.sim}(x, short = 0, round = NA, ...)
}
\arguments{
\item{x}{Output object from the simulations study program meerva.sim.block}

\item{short}{0 to produce extensive output summary, 1 to produce only a table of biases and MSEs}

\item{round}{number of decimal places to round to in some tables, NA for R default}

\item{...}{further arguments}
}
\value{
A summary print
}
\description{
Print Information for a meerva.sim Simulation Study Output Object (alias for summary.meerva.fit())
}
\seealso{
\code{\link{meerva.sim.block}} , \code{\link{summary.meerva.sim}}
}
