% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_210503.R
\name{reldif}
\alias{reldif}
\title{Calculate relative differences}
\usage{
reldif(a, b)
}
\arguments{
\item{a}{Object 1 for comparison}

\item{b}{Object 2 for comparison}
}
\value{
A object with relative differences
}
\description{
Calculate relative differences
}
