% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_210503.R
\name{subvec}
\alias{subvec}
\title{Subtract a vector from each row of a matrix}
\usage{
subvec(x, v)
}
\arguments{
\item{x}{A matrix}

\item{v}{A vector of length dim[x](2)}
}
\value{
A matrix
}
\description{
Subtract a vector from each row of a matrix
}
