% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210424.R
\name{summary.meerva}
\alias{summary.meerva}
\title{Summarize Information for a meerva Output Object}
\usage{
\method{summary}{meerva}(object, ...)
}
\arguments{
\item{object}{A meerva class object for summary.}

\item{...}{further arguments}
}
\value{
Summarize output
}
\description{
Summarize Information for a meerva Output Object
}
