% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210424.R
\name{ztest}
\alias{ztest}
\title{Ztest for beta coefficients}
\usage{
ztest(estimate, var, names, alpha = 0.05, round = NA)
}
\arguments{
\item{estimate}{beta estimates}

\item{var}{variance of estimates}

\item{names}{names of varaibles}

\item{alpha}{level for 91-alpha) confidence intervals}

\item{round}{number of decimal palces to rount some values}
}
\value{
table of summary statistics
}
\description{
Ztest for beta coefficients
}
