% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210424.R
\name{print.meerva}
\alias{print.meerva}
\title{Print Minimal Summary Information for a meerva Output Object}
\usage{
\method{print}{meerva}(x, ...)
}
\arguments{
\item{x}{A meerva class object for printing}

\item{...}{further arguments}
}
\value{
Print output
}
\description{
Print Minimal Summary Information for a meerva Output Object
}
