% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva_210503.R
\name{summary.meerva}
\alias{summary.meerva}
\title{Summarize Information for a meerva Output Object}
\usage{
\method{summary}{meerva}(object, alpha = 0.05, round = NA, ...)
}
\arguments{
\item{object}{A meerva class object for summary.}

\item{alpha}{level for (1-alpha) confidence intervals}

\item{round}{number of decimal places to print for some outputs}

\item{...}{further arguments}
}
\value{
Summarize output
}
\description{
Summarize Information for a meerva Output Object
}
