% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meerva.sim.tools_210503.R
\name{summary.meerva.sim}
\alias{summary.meerva.sim}
\title{Summarize Information for a meerva.sim Simulation Study Output Object}
\usage{
\method{summary}{meerva.sim}(object, short = 0, round = NA, ...)
}
\arguments{
\item{object}{Output object from the simulations study program meerva.sim.block}

\item{short}{0 to produce extensive output summary, 1 to produce only a table of biases and MSEs}

\item{round}{number of decimal places to round to in some tables, NA for R default}

\item{...}{further arguments}
}
\value{
A summary print
}
\description{
Summarize Information for a meerva.sim Simulation Study Output Object
}
\seealso{
\code{\link{meerva.sim.block}}
}
